use float/Py_ssize_t instead of double/long types in Falloffs

+++ animation_nodes/data_structures/falloffs/evaluation.pxd
@@ -2,4 +2,4 @@ cdef class FalloffEvaluator:
     '''The evaluator of a falloff is only valid while the falloff exists'''
     cdef object pyEvaluator
 
-    cdef double evaluate(self, void* value, long index)
+    cdef float evaluate(self, void *value, Py_ssize_t index)
