+++ animation_nodes/data_structures/falloffs/falloff_base.pxd
@@ -8,9 +8,9 @@ cdef class Falloff:
 
 cdef class BaseFalloff(Falloff):
     cdef str dataType
-    cdef double evaluate(self, void* object, long index)
+    cdef float evaluate(self, void *object, Py_ssize_t index)
 
 cdef class CompoundFalloff(Falloff):
     cdef list getDependencies(self)
     cdef list getClampingRequirements(self)
-    cdef double evaluate(self, double* dependencyResults)
+    cdef float evaluate(self, float *dependencyResults)
